web/requirements.txt
Django==2.0.5
django-filter==1.1.0
djangorestframework==3.8.2
gunicorn==19.8.1
Markdown==2.6.11
psycopg2==2.7.4
whitenoise==3.3.1
django-rest-auth[with_social]
